refactor: Split CustomUser and user resource code into a new accounts app
Creates a Django `accounts` app and moves identity / auth concerns
out of `ctj_api`. `ctj_api` now hosts only the recruitment-catalog
and taxonomy domain (Opportunity, Project, Role, Skill, SkillMatrix,
CommunityOfPractice). The new app earns its boundary by giving the
upcoming Stage 1 auth flow (signup, login, logout, me, csrf - lands
in a follow-up commit) a natural home and by isolating the eventual
Stage 2 work (PeopleDepot integration, Cognito ID-token mode) to a
single app surface. Done now while there's no real DB data to
migrate.

What moved (ctj_api -> accounts):

- `CustomUser` model.
- `user_detail` view + `UserDetailPermission` + `CustomUserReadSerializer`.
- `users/<uuid>/` URL route (still served at `/api/users/<uuid>/`;
  `accounts.urls` mounts at the same `/api/` prefix from `backend.urls`,
  ordered before `ctj_api.urls` so its catch-all doesn't shadow).
- `make_pm_user` / `make_regular_user` test factories.
- `test_users.py`.

Cross-app boundaries:

- `Opportunity.created_by` FK now uses `settings.AUTH_USER_MODEL`
  instead of importing `CustomUser` directly. Avoids circular
  imports (accounts.models needs `CommunityOfPractice` and
  `SkillMatrix` from `ctj_api`) and follows the Django-recommended
  pattern - the FK shape stays the same when Stage 2 swaps
  `AUTH_USER_MODEL`.
- `accounts.tests.common` imports its own factories;
  `ctj_api.tests.common` keeps domain factories;
  `test_opportunities.py` imports user factories from
  `accounts.tests.common`.

Settings:

- `INSTALLED_APPS` adds `accounts.apps.AccountsConfig` (before
  `ctj_api` for natural dependency order).
- `AUTH_USER_MODEL` flips from `"ctj_api.CustomUser"` to
  `"accounts.CustomUser"`.

Migration strategy:

- Nuked `ctj_api/migrations/0001_initial.py` and `0002_*.py`.
- Regenerated: `accounts/migrations/0001_initial.py` (CustomUser),
  `accounts/migrations/0002_initial.py` (FKs to ctj_api models),
  `ctj_api/migrations/0001_initial.py` (six domain models).
- Anyone with a local dev DB needs to nuke it (`make db-reset-hard`);
  no stage data exists.

Lint config:

- `.github/workflows/lint.yml`: add `accounts` to mypy and bandit
  invocations.
- `backend/pyproject.toml`: add `accounts/tests` to bandit
  `exclude_dirs`.
- `Makefile`: drop hardcoded `ctj_api.tests` from
  `local-test-backend` so test discovery picks up both apps.

Tests pass on a fresh test DB (10 tests across both apps).
