+++ backend/backend/urls.py
@@ -1,13 +1,18 @@
 """Top-level URL routing for the backend Django project.
 
-Three top-level mounts:
+Mounts:
 - `/admin/` for the Django admin site.
-- `/api/` delegates to `ctj_api.urls` for all CTJ API routes.
+- `/api/` delegates to `accounts.urls` then `ctj_api.urls` for the
+  full CTJ API surface.
 - Everything else falls through to the SPA catchall in
   `backend.views`, which serves an HTML template.
 
-Order matters: the SPA catchall is last so the explicit `/admin/`
-and `/api/` mounts win first.
+Order matters in two places:
+- The SPA catchall is last so the explicit `/admin/` and `/api/`
+  mounts win first.
+- `accounts.urls` is mounted BEFORE `ctj_api.urls` because
+  `ctj_api.urls` ends with a `/api/*` catch-all (`api_not_found`)
+  that would otherwise shadow accounts routes.
 
 Architectural note: the SPA catchall is a remnant from the
 pre-rewrite monolithic setup, when Django served both the API and
@@ -16,7 +21,7 @@ architecture runs the frontend as a separate Next.js container that
 handles all non-API traffic; anyone hitting the backend service
 directly at a non-`/admin/`, non-`/api/` path is probably misrouted.
 The catchall is preserved for now but is a candidate for removal -
-deferred out of this docs-only PR.
+deferred.
 """
 
 from django.contrib import admin
@@ -26,6 +31,7 @@ from . import views
 
 urlpatterns = [
     path("admin/", admin.site.urls),
+    path("api/", include("accounts.urls")),
     path("api/", include("ctj_api.urls"), name="api"),
     # Catch-all for frontend (React)
     re_path(
