+++ backend/ctj_api/admin.py
@@ -1,11 +1,12 @@
 """Django admin registrations for CTJ's domain models.
 
-All seven models are registered with the default `ModelAdmin` (no
-customization). Effect: admins see every field on every row, with
-no list filters, search fields, or read-only protections. This is
-enough for Stage 1 curation of the admin-managed reference tables
+All six domain models are registered with the default `ModelAdmin`
+(no customization). Effect: admins see every field on every row,
+with no list filters, search fields, or read-only protections.
+Enough for Stage 1 curation of the admin-managed reference tables
 (`Skill`, `Role`, `Project`, `CommunityOfPractice`) plus emergency
-edits to user records and opportunities.
+edits to opportunities. The `CustomUser` admin lives in
+`accounts.admin`.
 
 If a model's admin needs filters, list display, or search later,
 register it with a dedicated `ModelAdmin` subclass instead of the
@@ -21,7 +22,6 @@ from django.contrib import admin
 
 from .models import (
     CommunityOfPractice,
-    CustomUser,
     Opportunity,
     Project,
     Role,
@@ -34,5 +34,4 @@ admin.site.register(Role)
 admin.site.register(Skill)
 admin.site.register(Project)
 admin.site.register(SkillMatrix)
-admin.site.register(CustomUser)
 admin.site.register(Opportunity)
