+++ backend/ctj_api/apps.py
@@ -2,7 +2,12 @@ from django.apps import AppConfig
 
 
 class CtjApiConfig(AppConfig):
-    """Django app config for `ctj_api`, the project's only domain app.
+    """Django app config for `ctj_api`, the project's domain app.
+
+    Owns the recruitment-catalog and taxonomy models (Opportunity,
+    Project, Role, Skill, SkillMatrix, CommunityOfPractice) and
+    their views/serializers/permissions/urls. The `CustomUser`
+    identity model and the auth flow live in the `accounts` app.
 
     `default_auto_field` is set to `BigAutoField` to silence Django's
     startup warning about implicit auto field selection, but in
