+++ backend/ctj_api/models.py
@@ -1,10 +1,12 @@
 """Domain models for the CTJ platform.
 
-This module defines the seven Django models that back CTJ's core
+This module defines the six Django models that back CTJ's core
 flows: the practice-area taxonomy (`CommunityOfPractice`, `Role`),
-the skill system (`Skill`, `SkillMatrix`), the recruitment catalog
-(`Project`, `Opportunity`), and the user / identity model
-(`CustomUser`, which is the project's `AUTH_USER_MODEL`).
+the skill system (`Skill`, `SkillMatrix`), and the recruitment
+catalog (`Project`, `Opportunity`). The user / identity model
+(`CustomUser`) lives in the `accounts` app; FKs to it use
+`settings.AUTH_USER_MODEL` so this module avoids importing the
+user model directly.
 
 Stage 1 / Stage 2 boundary: several of these tables are locally
 curated in Stage 1 and migrate to PeopleDepot-backed shapes in
@@ -14,7 +16,7 @@ docs/developer/backend.md for the full migration plan.
 
 import uuid
 
-from django.contrib.auth.models import AbstractUser
+from django.conf import settings
 from django.db import models
 
 
@@ -236,87 +238,6 @@ class SkillMatrix(models.Model):
         return f"Skill matrix ID: {self.id}"
 
 
-class CustomUser(AbstractUser):
-    """
-    Summary:
-    - CTJ's user / identity model (the project's `AUTH_USER_MODEL`).
-
-    Business workflow:
-    - Created at signup via email + password (Stage 1) or via Cognito
-      (Stage 2).
-    - Identity reconciles to a PeopleDepot user record in Stage 2 via
-      the `people_depot_user_id` field; Stage 1 self-registered users
-      carry a `local:<uuid>` placeholder until that reconciliation.
-    - `isProjectManager` gates the ability to post opportunities in
-      the CMS (see `OpportunityPermission`).
-
-    Current policy:
-    - Subclasses Django's `AbstractUser` rather than extending the
-      default `User` via a OneToOne profile model. The trade-off is
-      that all user-shaped fields - identity, practice area, skills,
-      availability - live on a single table; reads are simpler
-      (`request.user.name` works directly, no join required), at the
-      cost of locking in the choice from day 1.
-    - `username` (inherited from `AbstractUser`) is set to the user's
-      email at signup time. That's what lets Django's default
-      `ModelBackend` resolve `authenticate(username=email,
-      password=...)` without writing a custom auth backend. Stage 2
-      (Cognito) replaces the password-based auth path entirely, so
-      this coupling is temporary.
-    - `people_depot_user_id` is unique-not-null; the `local:<uuid>`
-      placeholder satisfies the constraint until Stage 2 reconciliation.
-      The cleaner long-term fix is making the field nullable; deferred
-      to a schema migration.
-    - `isProjectManager` is camelCase intentionally; renaming would
-      require a schema migration and isn't worth the disruption solo.
-
-    Lifecycle control:
-    - `user-managed` (self-registration in Stage 1).
-
-    Visibility:
-    - `auth-required` via `/api/users/<uuid>/`; restricted further by
-      `UserDetailPermission`.
-    """
-
-    id = models.UUIDField(primary_key=True, default=uuid.uuid4, editable=False)
-    people_depot_user_id = models.CharField(max_length=255, unique=True)
-    name = models.CharField(max_length=100)
-    email = models.EmailField(unique=True)
-    community_of_practice = models.ForeignKey(
-        CommunityOfPractice,
-        on_delete=models.SET_NULL,
-        null=True,
-        blank=True,
-        related_name="user",
-    )
-    skills_learned_matrix = models.OneToOneField(
-        SkillMatrix,
-        on_delete=models.SET_NULL,
-        null=True,
-        blank=True,
-        related_name="user",
-        help_text="User's list of skills mapped to a mastery level (1-5).",
-    )
-    max_available_hours = models.IntegerField(
-        null=True, blank=True, help_text="User's available hours per week."
-    )
-    meeting_availability = models.JSONField(null=True, blank=True)
-    isProjectManager = models.BooleanField(
-        default=False,
-        help_text="A user that is a PM can create and edit opportunities in the CMS.",
-    )
-    created_at = models.DateTimeField(auto_now_add=True)
-    updated_at = models.DateTimeField(auto_now=True)
-
-    class Meta:
-        db_table = "users"
-        verbose_name = "User"
-        verbose_name_plural = "Users"
-
-    def __str__(self):
-        return self.name
-
-
 class Opportunity(models.Model):
     """
     Summary:
@@ -404,7 +325,7 @@ class Opportunity(models.Model):
         help_text="Status will determine how the opportunity will be shown publicly.",
     )
     created_by = models.ForeignKey(
-        CustomUser,
+        settings.AUTH_USER_MODEL,
         on_delete=models.SET_NULL,
         null=True,
         blank=True,
