+++ backend/ctj_api/views.py
@@ -1,10 +1,11 @@
 """DRF views for CTJ's domain endpoints, mounted under `/api/`.
 
 This module hosts the read/write endpoints for the recruitment
-catalog: opportunities, projects, roles, the skill catalog, the
-community-of-practice taxonomy, and per-user detail. It also
-provides the unauthenticated `healthcheck` endpoint and a
-JSON-shaped catch-all 404 handler for unknown `/api/*` paths.
+catalog: opportunities, projects, roles, the skill catalog, and
+the community-of-practice taxonomy. It also provides the
+unauthenticated `healthcheck` endpoint and a JSON-shaped catch-all
+404 handler for unknown `/api/*` paths. The per-user detail
+endpoint lives in `accounts.views`.
 
 Routing for these views lives in `ctj_api.urls` (mounted at
 `/api/` from `backend.urls`); permission classes live in
@@ -32,16 +33,14 @@ from rest_framework.response import Response
 
 from ctj_api.models import (
     CommunityOfPractice,
-    CustomUser,
     Opportunity,
     Project,
     Role,
     Skill,
 )
-from ctj_api.permissions import OpportunityPermission, UserDetailPermission
+from ctj_api.permissions import OpportunityPermission
 from ctj_api.serializers import (
     CommunityOfPracticeReadSerializer,
-    CustomUserReadSerializer,
     OpportunityReadSerializer,
     OpportunityWriteSerializer,
     ProjectReadSerializer,
@@ -128,42 +127,6 @@ def api_not_found(request, exception=None):
         },
         status=404,
     )
-
-
-@api_view(["GET"])
-@permission_classes([permissions.IsAuthenticated, UserDetailPermission])
-def user_detail(request, pk):
-    """
-    Summary:
-    - Return a single `CustomUser` record by UUID.
-    - Read-only - this view doesn't expose update or delete operations
-      on user records.
-    - Mutating user data goes through other paths (Django admin,
-      future Stage 2 PeopleDepot sync).
-
-    Flow:
-    1. Look up the `CustomUser` by primary-key UUID.
-    2. Trigger the object-level permission check (own-profile-only).
-    3. Serialize and return 200.
-
-    URL:
-    - GET /api/users/<uuid:pk>/
-
-    Auth:
-    - IsAuthenticated + UserDetailPermission
-
-    Errors:
-    - 403: Unauthenticated, OR `UserDetailPermission` denied
-      (requester is not the target user).
-    - 404: No user exists with the given UUID.
-    """
-    user = get_object_or_404(CustomUser, pk=pk)
-    # @permission_classes covers request-level (`has_permission`); the
-    # object-level (`has_object_permission`) check has to be triggered
-    # explicitly in FBVs since there's no APIView class to auto-call it.
-    request.parser_context["view"].check_object_permissions(request, user)
-    serializer = CustomUserReadSerializer(user)
-    return Response(serializer.data)
 
 
 class OpportunityViewSet(viewsets.ModelViewSet):
